Finished Lab 2

+++ Lab2/src/boardgame.c
@@ -2,7 +2,15 @@
 #include <stdlib.h>
 #include "boardgame.h"
 
-BoardGame* createBoardGame(char* name, float price, char* brandName){
+
+/*
+ * [createStore description]
+ * @param name static string of the name of the boardgame
+ * @param price double
+ * @param brandName static string of brand name
+ * @return dynamically created BoardGame
+*/
+BoardGame* createBoardGame(char* name, double price, char* brandName){
     BoardGame* bg = malloc(sizeof(BoardGame));
     bg->name = name;
     bg->price = price;
@@ -10,22 +18,38 @@ BoardGame* createBoardGame(char* name, float price, char* brandName){
     return bg;
 }
 
+/*
+ * Get name of board game
+ * @param bg BoardGame
+ * @return char* string of name
+*/
 char* getBGName(BoardGame *bg){
     return bg->name;
 }
 
-
-float getBGPrice(BoardGame *bg){
+/*
+ * Get price of board game
+ * @param bg BoardGame
+ * @return Double price
+*/
+double getBGPrice(BoardGame *bg){
     return bg->price;
 }
 
-
+/*
+ * Get brand name of board game
+ * @param bg BoardGame
+ * @return char* string of brand name
+*/
 char* getBrandName(BoardGame *bg){
     return bg->brandName;
 }
 
-
+/*
+ * Display the board game's name, price and brand name
+ * @param bg [BoardGame]
+*/  
 void printBoardGame(void *bg){
     bg = (BoardGame *) bg;
-    printf("%s: %.3f - %s\n", getBGName(bg), getBGPrice(bg), getBrandName(bg));
+    printf("%s: %.3lf - %s\n", getBGName(bg), getBGPrice(bg), getBrandName(bg));
 }
